Factor out server connection code from replay widget
Move methods that talk to the server outside the widget, to a separate
class. Note that the connection code here is almost identical to the
code in SecondaryServer - this needs some unification.
